refactor: PageResult分页字段从long改为int，避免Jackson序列化为String

+++ backend/src/main/java/com/linyi/pig/common/model/PageResult.java
@@ -18,22 +18,22 @@ public class PageResult<T> {
     /**
      * 总记录数
      */
-    private long total;
+    private int total;
 
     /**
      * 当前页码
      */
-    private long pageNum;
+    private int pageNum;
 
     /**
      * 每页记录数
      */
-    private long pageSize;
+    private int pageSize;
 
     /**
      * 总页数
      */
-    private long totalPages;
+    private int totalPages;
 
 
 }
